Legacy Merge (#48)
* Texture revisions in preparation for 0.4

* Sparks

In this commit I did the following:

- Fixed broken stuff that I didnt test cause I wanted it shipped early

This counts towards my final goodbye, invalidating it. Which is not nice. >:[

* Good Ending - A final word

In this commit I did the following:

- Fixed some stragglers.

This will be my real final commit of 0.4.0. Have fun, C+.

* Stair item models should be fixed now

* Bruh

In this commit I did the following:

- Devious seriously why you got to do it like this?

Guys no this is just DLC I swear please dont hurt me I dont mean it like I was scamming you.

* Weather Hackers

A final commit to wrap the remaining pieces up.

Co-authored-by: Carter <69820300+Cart3r1234@users.noreply.github.com>